TVGUI-1 Convert Home Page to TS
